Add some ENDPOINT
Please add an API with the help of django-rest-framework. How would you test the functionality?
Write an API endpoint to create the company along with office information
Write an simple read-only API endpoint for companies to get the company name + street+postal_code+city from the headquarter office
Write an simple read-only API endpoint to get all the offices for a company
Write an API endpoint to change the headquarter of the company